Revert ClassWriter aliasing
ClassWriter prefixes columns with names that correspond to Java reserved
words with '_'. For example, 'public' becomes '_public'. Later, when
we attempt to look up the column in the Hive schema via HCatalog, the
aliased column doesn't exist (assuming we've created a Hive schema that
exactly mirrors the schema of the source DB).

This patch coarsely removes prefixed underscores. A potential
improvement would be to check that the column name indeed is a Java
reserved word.